refactor: move TransactionResponse and parsing logic to a new parser file
